+++ src/Popover/Popover.tsx
@@ -1,17 +1,14 @@
 import React, {
   type FC,
   type ReactNode,
-  type HTMLProps,
   type PropsWithChildren,
   type ComponentProps,
   type MutableRefObject,
   type ReactElement,
   cloneElement,
-  forwardRef,
   useState,
   useCallback,
   useRef,
-  useEffect,
 } from "react";
 import { useMergeRefs } from "use-callback-ref";
 import {
@@ -28,13 +25,14 @@ import {
   autoUpdate,
 } from "@floating-ui/react";
 import cx from "classnames";
-import { Transition, TransitionStatus } from "react-transition-group";
+import { Transition } from "react-transition-group";
 
 import { useLockedBody } from "src/hooks/useLockedBody";
 import { useFloatingContext } from "src/FloatingProvider";
 
 import styles from "./Popover.module.scss";
 import { PopoverProvider, usePopoverContext } from "./context";
+import { GenericPopover } from "./GenericPopover";
 
 export { usePopoverContext };
 
@@ -71,66 +69,6 @@ const Overlay: FC<
     </Transition>
   );
 };
-
-export const GenericPopover = forwardRef<
-  HTMLDivElement,
-  HTMLProps<HTMLDivElement> & {
-    wrapperClassName?: string;
-    state: TransitionStatus;
-    offset?: Parameters<typeof offset>[0];
-  }
->(function GenericPopoverRefed(
-  {
-    className,
-    children,
-    wrapperClassName,
-    state,
-    offset: offsetOptions,
-    ...props
-  },
-  ref
-) {
-  const ownRef = useRef<HTMLDivElement>(null);
-  const merged = useMergeRefs([ref, ownRef]);
-
-  useEffect(() => {
-    const r = ownRef.current;
-
-    if (r && offsetOptions && typeof offsetOptions !== "function")
-      r.style.setProperty(
-        "--offset",
-        `${
-          typeof offsetOptions === "number"
-            ? offsetOptions
-            : offsetOptions.mainAxis ?? 4
-        }px`
-      );
-
-    return () => {
-      if (r) r.style.setProperty("--offset", "4px");
-    };
-  }, [offsetOptions]);
-
-  return (
-    <div
-      ref={merged}
-      className={cx(styles.wrapper, wrapperClassName, {
-        [styles.disappear]: state === "exiting" || state === "exited",
-      })}
-      {...props}
-    >
-      <div
-        className={cx(
-          "panel panel--bordered panel--raised",
-          styles.body,
-          className
-        )}
-      >
-        {children}
-      </div>
-    </div>
-  );
-});
 
 export type PopoverProps = PropsWithChildren<{
   onOpen?: () => unknown;