+++ src/Popover/index.ts
@@ -1,3 +1,4 @@
 export * from "./Popover";
 export * from "./PopoverTitle";
 export * from "./usePopover";
+export * from "./GenericPopover";